+++ apps/connect/release_notes.md
@@ -2,7 +2,7 @@
 
 ## upcoming
 
-* [new] Possibility to disable the plugin manager by setting environment variable FTRACK_DISABLE_PLUGIN_MANAGER to true.
+* [new] Possibility to disable the plugin manager by setting environment variable FTRACK_CONNECT_DISABLE_PLUGIN_MANAGER to true.
 
 
 ## 24.5.1